requirements.txt
tqdm
flickrapi
Pillow